backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pyjwt>=2.8.0
python-multipart>=0.0.6
google-genai>=1.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
pymongo>=4.6.0
numpy
